requirements.txt
absl-py==2.0.0
asgiref==3.7.2
asttokens==2.4.0
astunparse==1.6.3
attrs==23.1.0
backcall==0.2.0
biopandas==0.4.1
biopython==1.81
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cryptography==41.0.4
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.13
django-allauth==0.57.0
django-extensions==3.2.3
djongo==1.3.6
dnspython==2.4.2
executing==2.0.0
flatbuffers==23.5.26
fonttools==4.43.1
gast==0.5.4
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.59.0
h11==0.14.0
h5py==3.10.0
idna==3.4
ipython==8.16.1
ipywidgets==8.1.1
jedi==0.19.1
json2html==1.3.0
jupyterlab-widgets==3.0.9
keras==2.14.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
mongoengine==0.27.0
nglview==3.0.8
numpy==1.26.0
oauthlib==3.2.2
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.1
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.1
prompt-toolkit==3.0.39
protobuf==4.24.4
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
pymongo==3.12.3
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sortedcontainers==2.4.0
sqlparse==0.2.4
stack-data==0.6.3
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
traitlets==5.11.2
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
wcwidth==0.2.8
Werkzeug==3.0.0
widgetsnbextension==4.0.9
wrapt==1.14.1
wsproto==1.2.0
